Cleanup
Remove not used code in DynamicIndexStoreView, use entry sets iterators
use correct type for numberOfCollisions in EncodingIdMapper,
do proper non handled version check in RecordStorageCommandReaderFactory

+++ community/kernel/src/main/java/org/neo4j/unsafe/impl/batchimport/cache/idmapping/string/EncodingIdMapper.java
@@ -470,7 +470,7 @@ public class EncodingIdMapper implements IdMapper
         }
         workers.awaitAndThrowOnErrorStrict( RuntimeException.class );
 
-        int numberOfCollisions = 0;
+        long numberOfCollisions = 0;
         for ( DetectWorker detectWorker : workers )
         {
             numberOfCollisions += detectWorker.numberOfCollisions;
